feat(infra): Split the TypeScript type declaration
